+++ docs/CNAME
@@ -1,1 +1,1 @@
-bertok.qubitpi.org
+bertok.wilhelmlang.com
